refactor: prepare for multi-return values in GenerateForAnnotation
JsonSerializableGenerator now returns String instead of Future<String>
GeneratorHelper is now private and in the same file is the Generator
Encoding and decoding have been split into mixin classes
  These two aspects of the generator are nicely isolated hopefully
  making future fixes and additions easier

Many members of utils.dart have been moved closer to where they are used
  and made private when possible

Updated tests to be sync now that the core tested function is sync
