+++ json_serializable/lib/src/type_helper.dart
@@ -78,3 +78,9 @@ class UnsupportedTypeError extends Error {
 
   UnsupportedTypeError(this.type, this.expression, this.reason);
 }
+
+String commonNullPrefix(
+        bool nullable, String expression, String unsafeExpression) =>
+    nullable
+        ? '$expression == null ? null : $unsafeExpression'
+        : unsafeExpression;
