+++ json_serializable/lib/src/type_helpers/date_time_helper.dart
@@ -5,7 +5,6 @@
 import 'package:analyzer/dart/element/type.dart';
 import 'package:source_gen/source_gen.dart' show TypeChecker;
 import '../type_helper.dart';
-import '../utils.dart';
 
 class DateTimeHelper extends TypeHelper {
   const DateTimeHelper();
